requirements.txt
aiogram==3.3.0
aiohttp
requests
eng_to_ipa
python-dotenv